+++ src/accretion-geometry/intersections.jl
