+++ src/corona-to-disc/sky-geometry.jl
