+++ src/metrics/boyer-lindquist-fo.jl
@@ -1,6 +1,5 @@
 module __BoyerLindquistFO
-using ..FirstOrderMethods
-using ..DocStringExtensions
+using DocStringExtensions
 
 """
     $(TYPEDSIGNATURES)
@@ -355,21 +354,21 @@ end
 GradusBase.inner_radius(m::BoyerLindquistFO{T}) where {T} = m.M + √(m.M^2 - m.a^2)
 GradusBase.constrain(::BoyerLindquistFO{T}, u, v; μ = 0.0) where {T} = v[1]
 
-FirstOrderMethods.four_velocity(u, m::BoyerLindquistFO{T}, p) where {T} =
+four_velocity(u, m::BoyerLindquistFO{T}, p) where {T} =
     __BoyerLindquistFO.four_velocity(u, m.E, m.M, m.a, p)
-FirstOrderMethods.calc_lq(m::BoyerLindquistFO{T}, pos, vel) where {T} =
+calc_lq(m::BoyerLindquistFO{T}, pos, vel) where {T} =
     __BoyerLindquistFO.LQ(m.M, pos[2], m.a, pos[3], vel[3], vel[4])
 
-function FirstOrderMethods.Vr(m::BoyerLindquistFO{T}, u, p) where {T}
+function Vr(m::BoyerLindquistFO{T}, u, p) where {T}
     L, Q, _, _ = p
     __BoyerLindquistFO.Vr(m.E, L, m.M, Q, u[2], m.a)
 end
-function FirstOrderMethods.Vθ(m::BoyerLindquistFO{T}, u, p) where {T}
+function Vθ(m::BoyerLindquistFO{T}, u, p) where {T}
     L, Q, _, _ = p
     __BoyerLindquistFO.Vθ(m.E, L, Q, m.a, u[3])
 end
 
-function GeodesicTracer.alpha_beta_to_vel(m::BoyerLindquistFO{T}, u, α, β) where {T}
+function alpha_beta_to_vel(m::BoyerLindquistFO{T}, u, α, β) where {T}
     sinΦ, sinΨ = __BoyerLindquistFO.sinΦsinΨ(m.M, u[2], m.a, u[3], α, β)
     (β < 0.0 ? 1.0 : -1.0, sinΦ, sinΨ)
 end
