+++ src/metrics/johannsen-psaltis-ad.jl
@@ -43,6 +43,6 @@ $(FIELDS)
     ϵ3 = 0.0
 end
 
-GeodesicTracer.metric_components(m::JohannsenPsaltisAD{T}, rθ) where {T} =
+metric_components(m::JohannsenPsaltisAD{T}, rθ) where {T} =
     __JohannsenPsaltisAD.metric_components(m.M, m.a, m.ϵ3, rθ)
 GradusBase.inner_radius(m::JohannsenPsaltisAD{T}) where {T} = m.M + √(m.M^2 - m.a^2)
