+++ src/metrics/morris-thorne-ad.jl
@@ -29,6 +29,6 @@ $(FIELDS)
 end
 
 # implementation
-GeodesicTracer.metric_components(m::MorrisThorneAD{T}, rθ) where {T} =
+metric_components(m::MorrisThorneAD{T}, rθ) where {T} =
     __MorrisThorneAD.metric_components(m.b, rθ)
 GradusBase.inner_radius(m::MorrisThorneAD{T}) where {T} = 0.0
