+++ src/orbits/orbit-discovery.jl
@@ -1,4 +1,3 @@
-
 """
     $(TYPEDSIGNATURES)
